Move toFullTopicName to TopicName
